Bug fixed in in-memory push procedure using no sequence.

+++ StresSQL/dbo/Stored Procedures/usp_PushMessageImOltpNoSequence.sql
@@ -7,7 +7,7 @@ BEGIN ATOMIC
 
     DECLARE  @MessagePushed int
 	        ,@QueueSize     int = 200000
-			,@Slot          int;
+			,@Slot          bigint;
 
 	SET @MessagePushed = 0;
